+++ config/tasks.php
@@ -1,9 +1,9 @@
 <?php
-use Opulence\Applications\Tasks\Dispatchers\Dispatcher;
+use Opulence\Applications\Tasks\Dispatchers\TaskDispatcher;
 
 /**
  * ----------------------------------------------------------
  * Define the application task config
  * ----------------------------------------------------------
  */
-return new Dispatcher();
+return new TaskDispatcher();